+++ tests/unit/test_clearmot.py
@@ -4,15 +4,10 @@ import pytest
 from evaldet import MOTMetrics, Tracks
 
 
-def test_missing_frame_hyp():
+def test_missing_frame_hyp(missing_frame_pair):
     m = MOTMetrics()
 
-    gt = Tracks()
-    gt.add_frame(0, [0], np.array([[0, 0, 1, 1]]))
-    gt.add_frame(1, [0], np.array([[0, 0, 1, 1]]))
-
-    hyp = Tracks()
-    hyp.add_frame(0, [0], np.array([[0, 0, 1, 1]]))
+    gt, hyp = missing_frame_pair
     metrics = m.compute(
         gt, hyp, clearmot_metrics=True, id_metrics=False, hota_metrics=False
     )
@@ -22,15 +17,10 @@ def test_missing_frame_hyp():
     assert metrics["clearmot"]["IDSW"] == 0
 
 
-def test_missing_frame_gt():
+def test_missing_frame_gt(missing_frame_pair):
     m = MOTMetrics()
 
-    gt = Tracks()
-    gt.add_frame(1, [0], np.array([[0, 0, 1, 1]]))
-
-    hyp = Tracks()
-    hyp.add_frame(0, [0], np.array([[0, 0, 1, 1]]))
-    hyp.add_frame(1, [0], np.array([[0, 0, 1, 1]]))
+    hyp, gt = missing_frame_pair
     metrics = m.compute(
         gt, hyp, clearmot_metrics=True, id_metrics=False, hota_metrics=False
     )
@@ -43,11 +33,9 @@ def test_missing_frame_gt():
 def test_no_association_made():
     m = MOTMetrics()
 
-    gt = Tracks()
-    gt.add_frame(0, [0], np.array([[10, 10, 11, 11]]))
+    gt = Tracks(ids=[0], frame_nums=[0], detections=np.array([[10, 10, 11, 11]]))
+    hyp = Tracks(ids=[0], frame_nums=[0], detections=np.array([[0, 0, 1, 1]]))
 
-    hyp = Tracks()
-    hyp.add_frame(0, [0], np.array([[0, 0, 1, 1]]))
     metrics = m.compute(
         gt, hyp, clearmot_metrics=True, id_metrics=False, hota_metrics=False
     )
@@ -62,18 +50,17 @@ def test_no_association_made():
 @pytest.mark.parametrize("threshold", [0.3, 0.5, 0.7])
 def test_dist_threshold(threshold: float):
     m = MOTMetrics(clearmot_dist_threshold=threshold)
-    gt = Tracks()
-    gt.add_frame(
-        0,
-        [0, 1, 2, 3],
-        np.array([[0, 0, 1, 1], [0, 0, 1, 1], [0, 0, 1, 1], [0, 0, 1, 1]]),
+    gt = Tracks(
+        ids=[0, 1, 2, 3],
+        frame_nums=[0] * 4,
+        detections=np.array([[0, 0, 1, 1], [0, 0, 1, 1], [0, 0, 1, 1], [0, 0, 1, 1]]),
     )
-
-    hyp = Tracks()
-    hyp.add_frame(
-        0,
-        [0, 1, 2, 3],
-        np.array([[0, 0, 1, 0.2], [0, 0, 1, 0.4], [0, 0, 1, 0.6], [0, 0, 1, 0.8]]),
+    hyp = Tracks(
+        ids=[0, 1, 2, 3],
+        frame_nums=[0] * 4,
+        detections=np.array(
+            [[0, 0, 1, 0.2], [0, 0, 1, 0.4], [0, 0, 1, 0.6], [0, 0, 1, 0.8]]
+        ),
     )
 
     fn_res = {0.3: 3, 0.5: 2, 0.7: 1}
@@ -90,14 +77,12 @@ def test_sticky_association():
     """
     m = MOTMetrics()
 
-    gt = Tracks()
-    gt.add_frame(0, [0], np.array([[0, 0, 1, 1]]))
-    gt.add_frame(1, [0], np.array([[0, 0, 1, 1]]))
-
-    hyp = Tracks()
-    hyp.add_frame(0, [0], np.array([[0, 0, 1, 1]]))
-    hyp.add_frame(1, [0, 1], np.array([[0.1, 0.1, 1.1, 1.1], [0, 0, 1, 1]]))
-
+    gt = Tracks(ids=[0, 0], frame_nums=[0, 1], detections=np.array([[0, 0, 1, 1]] * 2))
+    hyp = Tracks(
+        ids=[0, 0, 1],
+        frame_nums=[0, 1, 1],
+        detections=np.array([[0, 0, 1, 1], [0.1, 0.1, 1.1, 1.1], [0, 0, 1, 1]]),
+    )
     metrics = m.compute(
         gt, hyp, clearmot_metrics=True, id_metrics=False, hota_metrics=False
     )
@@ -109,14 +94,8 @@ def test_sticky_association():
 def test_mismatch():
     m = MOTMetrics()
 
-    gt = Tracks()
-    gt.add_frame(0, [0], np.array([[0, 0, 1, 1]]))
-    gt.add_frame(1, [0], np.array([[0, 0, 1, 1]]))
-
-    hyp = Tracks()
-    hyp.add_frame(0, [0], np.array([[0, 0, 1, 1]]))
-    hyp.add_frame(1, [1], np.array([[0, 0, 1, 1]]))
-
+    gt = Tracks(ids=[0, 0], frame_nums=[0, 1], detections=np.array([[0, 0, 1, 1]] * 2))
+    hyp = Tracks(ids=[0, 1], frame_nums=[0, 1], detections=np.array([[0, 0, 1, 1]] * 2))
     metrics = m.compute(
         gt, hyp, clearmot_metrics=True, id_metrics=False, hota_metrics=False
     )
@@ -131,15 +110,10 @@ def test_persistent_mismatch():
     is not assigned."""
     m = MOTMetrics()
 
-    gt = Tracks()
-    gt.add_frame(0, [0], np.array([[0, 0, 1, 1]]))
-    gt.add_frame(1, [0], np.array([[0, 0, 1, 1]]))
-    gt.add_frame(2, [0], np.array([[0, 0, 1, 1]]))
-
-    hyp = Tracks()
-    hyp.add_frame(0, [0], np.array([[0, 0, 1, 1]]))
-    hyp.add_frame(2, [1], np.array([[0, 0, 1, 1]]))
-
+    gt = Tracks(
+        ids=[0] * 3, frame_nums=[0, 1, 2], detections=np.array([[0, 0, 1, 1]] * 3)
+    )
+    hyp = Tracks(ids=[0, 1], frame_nums=[0, 2], detections=np.array([[0, 0, 1, 1]] * 2))
     metrics = m.compute(
         gt, hyp, clearmot_metrics=True, id_metrics=False, hota_metrics=False
     )
@@ -148,20 +122,11 @@ def test_persistent_mismatch():
     assert metrics["clearmot"]["FP_CLEAR"] == 0
 
 
-def test_simple_case():
+def test_simple_case(simple_case):
     """Test a simple case with 3 frames and 2 detections/gts per frame."""
     m = MOTMetrics()
 
-    gt = Tracks()
-    gt.add_frame(0, [0, 1], np.array([[0, 0, 1, 1], [1, 1, 1, 1]]))
-    gt.add_frame(1, [0, 1], np.array([[0, 0, 1, 1], [2, 2, 1, 1]]))
-    gt.add_frame(2, [0, 1], np.array([[0, 0, 1, 1], [2, 2, 1, 1]]))
-
-    hyp = Tracks()
-    hyp.add_frame(0, [0, 1], np.array([[0, 0, 1, 1], [1, 1, 1, 1]]))
-    hyp.add_frame(1, [0, 1], np.array([[0.1, 0.1, 1, 1], [1, 1, 1, 1]]))
-    hyp.add_frame(2, [2, 1], np.array([[0.05, 0.05, 1, 1], [2, 2, 1, 1]]))
-
+    gt, hyp = simple_case
     metrics = m.compute(
         gt, hyp, clearmot_metrics=True, id_metrics=False, hota_metrics=False
     )