+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.3.0] - 2022-12-02
+
+### Added
+
+* Support for reading/writing to parquet formats ([#30](https://github.com/tadejsv/EvalDeT/pull/30))
+
+### Changed
+
+* Changed internal representation of attributes to use a single numpy array instead of per-frame dict ([#30](https://github.com/tadejsv/EvalDeT/pull/30))
 
 ## [0.2.1] - 2022-10-11